requirement.txt
pygame
azure-iot-device
azure-eventhub
azure-eventhub-checkpointstoreblob-aio
azure-iot-hub
pynput
keyboard